informationHandler/requirements.txt
fastapi[standard]
httpx
Pillow
requests